python/service-invocation-grpc/grpc-client/requirements.txt
dapr-dev >= 1.5.0rc1.dev
gunicorn
Flask